monitor: Separate Palette subcomponent into module
