+++ src/@types/monitor.d.ts
@@ -1,12 +1,4 @@
 /**
- * Interface for the Palette subcomponent proxied by the Monitor component.
- */
-export interface IPalette {
-    /** Returns a `Promise` for the list of palette sections. */
-    getSections: () => Promise<string[]>;
-}
-
-/**
  * Interface for the Artboard subcomponent by the Monitor component.
  */
 export interface IArtboard {
