+++ src/views/palette/Palette.tsx
@@ -2,13 +2,13 @@
 
 import { IPaletteProps } from '../../@types/palette';
 
+// -- other components -----------------------------------------------------------------------------
+
+import Section from './Section';
+
 // -- stylesheet -----------------------------------------------------------------------------------
 
 import './Palette.scss';
-
-// -- Views ----------------------------------------------------------------------------------------
-
-import Section from './Section';
 
 // -- view component definition --------------------------------------------------------------------
