+++ src/views/palette/PopUp.tsx
@@ -1,11 +1,11 @@
 // -- types ----------------------------------------------------------------------------------------
+
 import { IPopUpBlocks } from '../../@types/palette';
+
 // -- stylesheet -----------------------------------------------------------------------------------
 
 import './PopUp.scss';
 
-// -- Views ----------------------------------------------------------------------------------------
-import { useState } from 'react';
 // -- view component definition --------------------------------------------------------------------
 
 /**
